third_party/blink/web_tests/platform/generic/svg/animations/additive-from-to-fill-animation-expected.txt

PASS getComputedStyle(document.querySelector('rect')).fill is 'rgb(0, 128, 0)'
PASS successfullyParsed is true

TEST COMPLETE
